feat: add additional, optional tags parameter to set-datadog-team
Add an input parameter to set-datadog-team to allow additional tags to
be set along with the team for the current top-level workflow in
Datadog.

This will allow workflows to add additional tags that can be
used by Datadog resources to refine the queries they use to identify
specific GitHub Actions workflows.

For example, the orchestra e2e test run for a server SDK release can
add a tag like `release-version:v6.7.0` and then a specific Datadog
monitor can be triggered when that test run completes.